Rework on FindMusic method, replay the message signature at the end and enhance logging
Refactored the `FindMusic` method to accept a semicolon separated list of files for searching, instead of a single one.
Introduced a variable, `signaturePlaylist`, to store audio files associated with a message's signature. After all body audio files are played, the signature audio files are replayed.
Enhanced logging in `PlayAndWait` and `FindMusic`.

Commented out the last part of the `FindMusic` method which searched for the last part of the `toFind` string in all root directories.